Updated Docs / Moved Examples
This patch updates the documentation to reflect how to use the Makefile
to run the benchmarks and examples. This update also renames the
examples so they start at 1 instead of 0.
