+++ examples/02/main.go
@@ -5,76 +5,37 @@ import (
 
 	log "github.com/emccode/gournal"
 	glogrus "github.com/emccode/gournal/logrus"
+	gzap "github.com/emccode/gournal/zap"
 )
 
 func main() {
+	// Make a Zap-based Appender the default appender for when one is not
+	// present in a Context, or when a nill Context is provided to a logging
+	// function.
+	log.DefaultAppender = gzap.New()
+
+	// The following call fails to provide a valid Context argument. In this
+	// case the DefaultAppender is used.
+	log.WithFields(map[string]interface{}{
+		"size":     2,
+		"location": "Boston",
+	}).Error(nil, "Hello %s", "Bob")
+
 	ctx := context.Background()
 	ctx = context.WithValue(ctx, log.LevelKey, log.InfoLevel)
+
+	// Even though this next call provides a valid Context, there is no
+	// Appender present in the Context so the DefaultAppender will be used.
+	log.Info(ctx, "Hello %s", "Mary")
+
 	ctx = context.WithValue(ctx, log.AppenderKey, glogrus.New())
 
-	ctx = context.WithValue(
-		ctx,
-		log.FieldsKey,
-		map[string]interface{}{
-			"name":  "Venus",
-			"color": 0x00ff00,
-		})
-
-	// The following log entry will print the message and the name and color
-	// of the planet.
-	log.Info(ctx, "Discovered planet")
-
-	ctx = context.WithValue(
-		ctx,
-		log.FieldsKey,
-		func() map[string]interface{} {
-			return map[string]interface{}{
-				"galaxy":   "Milky Way",
-				"distance": 42,
-			}
-		})
-
-	// The following log entry will print the message and the galactic location
-	// and distance of the planet.
-	log.Info(ctx, "Discovered planet")
-
-	// Create a Context with the FieldsKey that points to a function which
-	// returns a Context's derived fields based upon what data was provided
-	// to a the log function.
-	ctx = context.WithValue(
-		ctx,
-		log.FieldsKey,
-		func(ctx context.Context,
-			lvl log.Level,
-			fields map[string]interface{},
-			args ...interface{}) map[string]interface{} {
-
-			if v, ok := fields["z-value"].(int); ok {
-				delete(fields, "z-value")
-				return map[string]interface{}{
-					"point": struct {
-						x int
-						y int
-						z int
-					}{1, -1, v},
-				}
-			}
-
-			return map[string]interface{}{
-				"point": struct {
-					x int
-					y int
-				}{1, -1},
-			}
-		})
-
-	// The following log entry will print the message and two-dimensional
-	// location information about the planet.
-	log.Info(ctx, "Discovered planet")
-
-	// This log entry, however, will print the message and the same location
-	// information, however, because the function used to derive the Context's
-	// fields inspects the field's "z-value" key, it will add that data to the
-	// location information, making it three-dimensional.
-	log.WithField("z-value", 3).Info(ctx, "Discovered planet")
+	// This last log function uses a Context that has been created with a
+	// Logrus Appender. Even though the DefaultAppender is assigned and is a
+	// Zap-based logger, this call will utilize the Context Appender instance,
+	// a Logrus Appender.
+	log.WithFields(map[string]interface{}{
+		"size":     1,
+		"location": "Austin",
+	}).Warn(ctx, "Hello %s", "Alice")
 }
